// src/types/theme.ts
export interface ThemeWeights {
    text: TextStyle['fontWeight'];
    h1?: TextStyle['fontWeight'];
    h2?: TextStyle['fontWeight'];
    h3?: TextStyle['fontWeight'];
    h4?: TextStyle['fontWeight'];
    h5?: TextStyle['fontWeight'];
    p?: TextStyle['fontWeight'];

    thin: TextStyle['fontWeight'];
    extralight: TextStyle['fontWeight'];
    light: TextStyle['fontWeight'];
    normal: TextStyle['fontWeight'];
    medium: TextStyle['fontWeight'];
    semibold?: TextStyle['fontWeight'];
    bold?: TextStyle['fontWeight'];
    extrabold?: TextStyle['fontWeight'];
    black?: TextStyle['fontWeight'];
}

export interface ThemeFonts {
    text: string;
    h1: string;
    h2: string;
    h3: string;
    h4: string;
    h5: string;
    p: string;
    thin: string;
    extralight: string;
    light: string;
    normal: string;
    medium: string;
    bold: string;
    semibold: string;
    extrabold: string;
    black: string;
}

export interface ThemeLineHeights {
    text: number;
    h1: number;
    h2: number;
    h3: number;
    h4: number;
    h5: number;
    p: number;
}